chore: refactoring file-type to new structure
